09/09/2024

+++ data/doctors.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { Doctor, User } = require('../models/index')
+const { Doctor, User, Counter } = require('../models/index')
 const router = express.Router();
 const { Op } = require('sequelize')
 const bcrypt = require("bcryptjs")
@@ -13,7 +13,7 @@ const io = socketIo(server);
 router.post('/create_doctor', async (req, res) => {
     const { name, phone, service,room } = req.body;
     let newPass = await bcrypt.hash(phone,6)
-    console.log("creating a doctor",phone)
+    let role
     try {
         if(name.trim() === ''){
             return res.status(400).json({ error: 'name is required' });
@@ -40,24 +40,17 @@ router.post('/create_doctor', async (req, res) => {
                     phone,
                     service,
                     room,
-                    role:"doctor",
+                    role: service==="meds"?"medical_recorder":service==="accounts"?"accountant":service==="payment"?"cashier":"doctor",
                     password: newPass
                 })
                 await User.create({
                     name,
                     phone,
-                    role:"doctor",
+                    role:service==="meds"?"medical_recorder":service==="accounts"?"accountant":service==="payment"?"cashier":"doctor",
                     password:newPass,
                     service,
-                    room
+                    counter: room
                 })
-                io.on("connection",(socket)=> {
-                    console.log('emitting diode please wait...')
-                    io.emit('data',newAtt)
-                    // socket.on("disconnect", ()=> {
-                    //   console.log("websocket disconnected")
-                    // })
-                  })
                 res.json(newAtt)
             }
         }
@@ -101,9 +94,16 @@ router.put('/delete_doctor/:id', async (req, res) => {
         const service = await Doctor.findByPk(id);
         if (!service) {
         return res.status(404).json({ error: 'doctor not found' });
+        }else{
+            const user = await User.findOne({
+                where: {phone: service.phone}
+            })
+            if(user){
+                await user.destroy()
+                await service.destroy();
+                res.status(204).end();
+            }
         }
-        await service.destroy();
-        res.status(204).end();
     } catch (err) {
         //next({error: err})
         res.status(500).json({ error: err });
@@ -124,6 +124,71 @@ router.put('/edit_counter/:id', async (req, res) => {
             await service.update(newData);
             //res.status(204).end();
             res.json(service);   
+        }
+    } catch (err) {
+        //next({error: err})
+        res.status(500).json({ error: err });
+    }
+});
+router.put('/edit_doctor/:id', async (req, res) => {
+    const id = req.params.id
+    const newData = req.body
+    try {
+        if(!newData.fields.name.trim()===""){
+            return res.status(404).json({ error: 'name is empty' }); 
+        }else{
+            const service = await Doctor.findByPk(id);
+            if (!service) {
+            return res.status(404).json({ error: 'service not found' });
+            }else{
+                const user = await User.findOne({
+                    where: {phone: service.phone}
+                });
+                const counter = await Counter.findOne({
+                    where: {name: newData.fields.service}
+                });
+                if(!user || ! counter){
+                    return res.status(404).json({ error: 'user not found' });  
+                }else{
+                    const pass = await bcrypt.hash(newData.fields.phone,6)
+                    if(newData.fields.phone === user.phone){
+                        await service.update({
+                            name: newData.fields.name,
+                            phone: newData.fields.phone,
+                            service: newData.fields.service,
+                            room: newData.fields.service !=="clinic"?counter.namba:newData.fields.counter,
+                            role:newData.fields.service==="meds"?"medical_recorder":newData.fields.service==="accounts"?"accountant":newData.fields.service==="payment"?"cashier":"doctor",
+                        });
+                        await user.update({
+                            name: newData.fields.name,
+                            phone: newData.fields.phone,
+                            service: newData.fields.service,
+                            counter: newData.fields.service !=="clinic"?counter.namba:newData.fields.counter,
+                            role:newData.fields.service==="meds"?"medical_recorder":newData.fields.service==="accounts"?"accountant":newData.fields.service==="payment"?"cashier":"doctor",
+                        });
+                        res.json(service);
+                    }else{
+                        await service.update({
+                            name: newData.fields.name,
+                            phone: newData.fields.phone,
+                            service: newData.fields.service,
+                            room: newData.fields.service !=="clinic"?counter.namba:newData.fields.counter,
+                            password: pass,
+                            role:newData.fields.service==="meds"?"medical_recorder":newData.fields.service==="accounts"?"accountant":newData.fields.service==="payment"?"cashier":"doctor",
+                        });
+                        await user.update({
+                            name: newData.fields.name,
+                            phone: newData.fields.phone,
+                            service: newData.fields.service,
+                            counter: newData.fields.service !=="clinic"?counter.namba:newData.fields.counter,
+                            password: pass,
+                            role:newData.fields.service==="meds"?"medical_recorder":newData.fields.service==="accounts"?"accountant":newData.fields.service==="payment"?"cashier":"doctor",
+                        });
+                        //res.status(204).end();
+                        res.json(service);
+                    }
+                }
+            }   
         }
     } catch (err) {
         //next({error: err})
